+++ functions.php
@@ -60,37 +60,9 @@ if ( version_compare( PHP_VERSION, '5.3.29' ) < 0 ) {
 	return;
 }
 
+require_once 'start.php';
+
 require_once 'globals/utilities.php';
 require_once 'globals/hooks.php';
 require_once 'globals/sanitize-functions.php';
 
-/**
- * Run theme functionality
- */
-function neve_run() {
-	$vendor_file = trailingslashit( get_template_directory() ) . 'vendor/autoload.php';
-	if ( is_readable( $vendor_file ) ) {
-		require_once $vendor_file;
-	}
-
-	require_once 'autoloader.php';
-	$autoloader_class = '\\Neve\\Autoloader';
-	$autoloader       = new $autoloader_class;
-	$autoloader->add_namespace( 'Neve', get_template_directory() . '/inc/' );
-	if ( defined( 'NEVE_PRO_SPL_ROOT' ) ) {
-		$autoloader->add_namespace( 'Neve_Pro', NEVE_PRO_SPL_ROOT );
-	}
-	$autoloader->register();
-
-	$core_loader_class = '\\Neve\\Core\\Core_Loader';
-	if ( class_exists( $core_loader_class ) ) {
-		new $core_loader_class;
-	}
-
-	$addon_loader_class = '\\Neve_Pro\\Core\\Loader';
-	if ( class_exists( $addon_loader_class ) ) {
-		$addon_loader_class::instance();
-	}
-}
-
-neve_run();
